backend/requirements.txt
# Django Core
Django>=5.2.6
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8

# Image Handling
cloudinary>=1.36.0
Pillow>=10.0.0

# Database
psycopg2-binary>=2.9.9  # For PostgreSQL (optional)

# HTTP Requests (for WhatsApp service integration)
requests>=2.31.0

# Utilities
python-dateutil>=2.8.2

# Development
ipython>=8.18.0  # Optional, for better Django shell

